libindex: fix manifest_index unique constraint
this commit adds a migration to fix the missing
unique constraint on the manifest_index table.

without this constraint the manifest_index table can
bloat to absurd size.

in order to deal with tables which have already bloated
to absurd size we simply truncate the manifest_index table.

we also truncate the manifest_scanned table which allows
all subsequent manifests submitted to clair to be re-indexed.

this re-index will re-build the manifest_index for the manifest
being submitted, but perform no other work since all layer data
still exists for a previously seen manifest.

Signed-off-by: ldelossa <ldelossa@redhat.com>